pushing back DisableTxOffload into types.go
