+++ types/types.go
@@ -8,7 +8,10 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/containernetworking/plugins/pkg/ns"
 	"github.com/docker/go-connections/nat"
+	log "github.com/sirupsen/logrus"
+	"github.com/srl-labs/containerlab/utils"
 )
 
 // Link is a struct that contains the information of a link between 2 containers.
@@ -133,6 +136,27 @@ type NodeConfig struct {
 	Extras  *Extras    `json:"extras,omitempty"`
 	WaitFor []string   `json:"wait-for,omitempty"`
 	DNS     *DNSConfig `json:"dns,omitempty"`
+}
+
+func DisableTxOffload(n *NodeConfig) error {
+	// skip this if node runs in host mode
+	if strings.ToLower(n.NetworkMode) == "host" {
+		return nil
+	}
+	// disable tx checksum offload for linux containers on eth0 interfaces
+	nodeNS, err := ns.GetNS(n.NSPath)
+	if err != nil {
+		return err
+	}
+	err = nodeNS.Do(func(_ ns.NetNS) error {
+		// disabling offload on eth0 interface
+		err := utils.EthtoolTXOff("eth0")
+		if err != nil {
+			log.Infof("Failed to disable TX checksum offload for 'eth0' interface for Linux '%s' node: %v", n.ShortName, err)
+		}
+		return err
+	})
+	return err
 }
 
 type GenericFilter struct {
